dashboard/requirements.txt
Flask==3.1.3
pyntcore==2024.3.1.0
